requirements.txt
flask>=3.0.0
pymupdf>=1.24.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
